SBERDOMA-62 user.User move testSchema
