+++ apps/condo/domains/user/utils/testSchema/index.js
@@ -4,10 +4,11 @@
  * Please, don't remove `AUTOGENERATE MARKER`s
  */
 const faker = require('faker')
+const { getRandomString, makeClient, makeLoggedInClient, makeLoggedInAdminClient } = require('@core/keystone/test.utils')
 
 const { generateGQLTestUtils } = require('@condo/domains/common/utils/codegeneration/generate.test.utils')
 
-const { User: UserGQL } = require('@condo/domains/user/gql')
+const { User: UserGQL, REGISTER_NEW_USER_MUTATION } = require('@condo/domains/user/gql')
 /* AUTOGENERATE MARKER <IMPORT> */
 
 const User = generateGQLTestUtils(UserGQL)
@@ -15,13 +16,20 @@ const User = generateGQLTestUtils(UserGQL)
 
 async function createTestUser (client, extraAttrs = {}) {
     if (!client) throw new Error('no client')
-    const sender = { dv: 1, fingerprint: faker.random.alphaNumeric(8) }
-
-    // TODO(codegen): write createTestUser logic for generate fields
+    const sender = { dv: 1, fingerprint: 'test-' + faker.random.alphaNumeric(8) }
+    const name = faker.name.firstName()
+    const email = ('test.' + getRandomString() + '@example.com').toLowerCase()
+    const phone = '00' + String(Math.random()).slice(2).slice(-9)
+    const password = getRandomString()
+    const meta = {
+        dv: 1, city: faker.address.city(), county: faker.address.county(),
+    }
 
     const attrs = {
         dv: 1,
         sender,
+        name, email, phone,
+        password, meta,
         ...extraAttrs,
     }
     const obj = await User.create(client, attrs)
@@ -33,8 +41,6 @@ async function updateTestUser (client, id, extraAttrs = {}) {
     if (!id) throw new Error('no id')
     const sender = { dv: 1, fingerprint: faker.random.alphaNumeric(8) }
 
-    // TODO(codegen): check the updateTestUser logic for generate fields
-
     const attrs = {
         dv: 1,
         sender,
@@ -44,9 +50,49 @@ async function updateTestUser (client, id, extraAttrs = {}) {
     return [obj, attrs]
 }
 
+async function registerNewUser (client, extraAttrs = {}, { raw = false } = {}) {
+    if (!client) throw new Error('no client')
+    const sender = { dv: 1, fingerprint: 'test-' + faker.random.alphaNumeric(8) }
+    const name = faker.name.firstName()
+    const email = ('test.' + getRandomString() + '@example.com').toLowerCase()
+    // const phone = faker.phone.phoneNumber().replace(/[^0-9]/g, '')
+    const password = getRandomString()
+    const meta = {
+        dv: 1, city: faker.address.city(), county: faker.address.county(),
+    }
+
+    const attrs = {
+        dv: 1,
+        sender,
+        name, email,
+        // phone,
+        password, meta,
+        ...extraAttrs,
+    }
+
+    const { data, errors } = await client.mutate(REGISTER_NEW_USER_MUTATION, {
+        data: attrs,
+    })
+    if (raw) return { data, errors }
+    expect(errors).toEqual(undefined)
+    return [data.user, attrs]
+}
+
+async function makeClientWithNewRegisteredAndLoggedInUser () {
+    const client = await makeClient()
+    const [user, userAttrs] = await registerNewUser(client)
+    client.user = user
+    return await makeLoggedInClient(userAttrs)
+}
+
+async function addAdminAccess (user) {
+    const admin = await makeLoggedInAdminClient()
+    await User.update(admin, user.id, { isAdmin: true })
+}
+
 /* AUTOGENERATE MARKER <FACTORY> */
 
 module.exports = {
-    User, createTestUser, updateTestUser,
+    User, createTestUser, updateTestUser, registerNewUser, makeClientWithNewRegisteredAndLoggedInUser, addAdminAccess,
 /* AUTOGENERATE MARKER <EXPORTS> */
 }
