+++ apps/condo/utils/testSchema/Organization.js
@@ -1,5 +1,5 @@
 const faker = require('faker')
-const { makeClientWithNewRegisteredAndLoggedInUser } = require('./User')
+const { makeClientWithNewRegisteredAndLoggedInUser } = require('@condo/domains/user/utils/testSchema')
 const { Organization, OrganizationEmployee, REGISTER_NEW_ORGANIZATION_MUTATION, ACCEPT_OR_REJECT_ORGANIZATION_INVITE_BY_ID_MUTATION, INVITE_NEW_ORGANIZATION_EMPLOYEE_MUTATION } = require('../../gql/Organization')
 
 async function createOrganization (client, extraAttrs = {}) {
